[#11680] refactor(idp-basic): Extract shared System.exit test helper
Deduplicate SecurityManager-based exit interception used by Basic auth
and IdP REST feature configuration tests.
