+++ plugins/idp-basic/src/test/java/org/apache/gravitino/idp/web/rest/feature/TestIdpRESTFeature.java
@@ -27,6 +27,8 @@ import com.google.common.collect.Lists;
 import org.apache.gravitino.Config;
 import org.apache.gravitino.Configs;
 import org.apache.gravitino.auth.AuthenticatorType;
+import org.apache.gravitino.idp.SystemExitTestHelper;
+import org.apache.gravitino.idp.SystemExitTestHelper.SystemExitException;
 import org.junit.jupiter.api.Test;
 
 class TestIdpRESTFeature {
@@ -41,7 +43,11 @@ class TestIdpRESTFeature {
             AuthenticatorType.BASIC.name().toLowerCase()));
 
     SystemExitException exception =
-        assertThrows(SystemExitException.class, () -> validateWithExitGuard(config));
+        assertThrows(
+            SystemExitException.class,
+            () ->
+                SystemExitTestHelper.runWithExitGuard(
+                    () -> IdpRESTFeature.validateConfiguration(config)));
 
     assertEquals(1, exception.status());
   }
@@ -53,7 +59,11 @@ class TestIdpRESTFeature {
         Configs.AUTHENTICATORS, Lists.newArrayList(AuthenticatorType.OAUTH.name().toLowerCase()));
 
     SystemExitException exception =
-        assertThrows(SystemExitException.class, () -> validateWithExitGuard(config));
+        assertThrows(
+            SystemExitException.class,
+            () ->
+                SystemExitTestHelper.runWithExitGuard(
+                    () -> IdpRESTFeature.validateConfiguration(config)));
 
     assertEquals(1, exception.status());
   }
@@ -65,40 +75,5 @@ class TestIdpRESTFeature {
         Configs.AUTHENTICATORS, Lists.newArrayList(AuthenticatorType.BASIC.name().toLowerCase()));
 
     assertDoesNotThrow(() -> IdpRESTFeature.validateConfiguration(config));
-  }
-
-  @SuppressWarnings("removal")
-  private static void validateWithExitGuard(Config config) {
-    SecurityManager original = System.getSecurityManager();
-    System.setSecurityManager(
-        new SecurityManager() {
-          @Override
-          public void checkExit(int status) {
-            throw new SystemExitException(status);
-          }
-
-          @Override
-          public void checkPermission(java.security.Permission perm) {
-            // Allow test execution.
-          }
-        });
-    try {
-      IdpRESTFeature.validateConfiguration(config);
-    } finally {
-      System.setSecurityManager(original);
-    }
-  }
-
-  private static final class SystemExitException extends SecurityException {
-    private final int status;
-
-    private SystemExitException(int status) {
-      super("System.exit(" + status + ")");
-      this.status = status;
-    }
-
-    private int status() {
-      return status;
-    }
   }
 }
